♻️ Refactor: 컴포넌트 위치 변경
refactor: 컴포넌트 위치 변경
