+++ src/pages/QuestionForm/components/bottomSheet/QuestionTitleBottomSheet.tsx
@@ -1,10 +1,10 @@
 import { BottomSheet, TextField } from "@toss/tds-mobile";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { QUESTION_TYPE_ROUTES } from "../../../constants/routes";
-import { useSurvey } from "../../../contexts/SurveyContext";
-import type { QuestionType } from "../../../types/survey";
-import { createQuestion } from "../../../utils/questionFactory";
+import { QUESTION_TYPE_ROUTES } from "../../../../constants/routes";
+import { useSurvey } from "../../../../contexts/SurveyContext";
+import type { QuestionType } from "../../../../types/survey";
+import { createQuestion } from "../../../../utils/questionFactory";
 
 interface QuestionTitleBottomSheetProps {
 	onClose: () => void;
